feedback: Use pick delegate
