+++ backend/bot.js
@@ -1,27 +1,20 @@
-const {sample, pull, pullAll} = require("lodash");
+const {sample, pull} = require("lodash");
 const Player = require("./player");
 
 module.exports = class extends Player {
-  constructor(isDecadent) {
+  constructor() {
     super({
       isBot: true,
       isConnected: true,
       name: "bot",
-      id: "",
-      isDecadent: isDecadent
+      id: ""
     });
   }
 
   getPack(pack) {
     const randomPick = sample(pack);
     this.picks.push(randomPick.name);
-    
-    if (!this.isDecadent) {
-      pull(pack, randomPick);
-    } else {
-      pullAll(pack, pack);
-    }
-    
+    pull(pack, randomPick);
     this.emit("pass", pack);
   }
 };
